backend/requirements.txt
tensorflow
keras
numpy
opencv-python
matplotlib
flask
pillow
flask-cors
flask-sqlalchemy
flask-bcrypt
flask-jwt-extended
python-dotenv
pytest
